Render many more storable instances
We now render the structs in dependency order, keeping track of which ones have
storable instances. This allows us to use `poke` instead of `pokeCStruct` much
more often, which in turn causes the writes to be in IO instead of `ContT IO`
which is the criterion for a Storable instance.

Closes https://github.com/expipiplus1/vulkan/issues/209