Fix forward type family in boot file (9.8.2)
Fixes an hs-boot file which had an incorrect forward declaration of a
closed type family. vulkan would successfully compile despite the
hs-boot error with 9.8.1, but not with 9.8.2 because of bug fixes which
makes the compiler now reject this invalid program.

We need to use the weird `type family A x where ..` (with two literal
dots after `where`) to declare the type family in the boot file.

Fixes compilation with 9.8.2